Rewrite what I have so far as a single struct.
Before I was approaching this by creating a trait that would be
implented by a `Plane` struct and a `Sphere` struct. This ended up
getting to a point were I was trying to figure out if I needed dynamic
dispatch and it felt like I had over complicated things.

Instead I am heading in the direction of a `Shape` struct that has a
`shape_type` member. This means that each of the method for the shape
struct will be able to use different behaviour by matching on the
`shape_type` of `self`.
