+++ src/lib.rs
@@ -5,10 +5,8 @@ pub mod lighting;
 pub mod lights;
 pub mod material;
 pub mod matrix;
-pub mod plane;
 pub mod ray;
 pub mod shape;
-pub mod sphere;
 pub mod transformation;
 pub mod tuple;
 pub mod world;
