--- src/sphere.rs
@@ -1,162 +1,0 @@
-use crate::material;
-use crate::matrix;
-use crate::matrix::{Inverse, Transpose};
-use crate::ray;
-use crate::shape::Shape;
-use crate::tuple;
-
-#[derive(Debug, PartialEq)]
-pub struct Sphere {
-    transform: matrix::Matrix4,
-    pub material: material::Material,
-}
-
-impl Shape for Sphere {
-    fn default() -> Self {
-        return Sphere {
-            transform: matrix::Matrix4::IDENTITY,
-            material: material::material(),
-        };
-    }
-
-    fn transformation_matrix(&self) -> &matrix::Matrix4 {
-        &self.transform
-    }
-
-    fn set_transformation_matrix(&mut self, new_transform: matrix::Matrix4) {
-        self.transform = new_transform;
-    }
-
-    fn normal_at(&self, world_point: tuple::Point) -> tuple::Vector {
-        let transform_inverse = self.transform.inverse().unwrap();
-        let object_point = transform_inverse * world_point;
-        let object_normal = object_point - tuple::Point::new(0.0, 0.0, 0.0);
-        let mut world_normal = transform_inverse.transpose() * object_normal;
-        // This is sorta a cheat to skip finding the submatrix.
-        world_normal.w = 0.0;
-        return tuple::normalize(&world_normal);
-    }
-
-    fn local_intersect(&self, _local_ray: ray::Ray) -> tuple::Point {
-        tuple::Point::new(0.0, 1.0, 0.0)
-    }
-}
-
-#[cfg(test)]
-mod sphere_tests {
-    use crate::assert_tuple_approx_eq;
-    use crate::material;
-    use crate::matrix;
-    use crate::shape::Shape;
-    use crate::sphere;
-    use crate::transformation::Transform;
-    use crate::tuple;
-
-    #[test]
-    fn test_normal_on_a_sphere_on_the_x() {
-        let sphere = sphere::Sphere::default();
-
-        let normal = sphere.normal_at(tuple::Point::new(1.0, 0.0, 0.0));
-
-        let expected = tuple::Vector::new(1.0, 0.0, 0.0);
-        assert_eq!(expected, normal);
-    }
-
-    #[test]
-    fn test_normal_on_a_sphere_on_the_y() {
-        let sphere = sphere::Sphere::default();
-
-        let normal = sphere.normal_at(tuple::Point::new(0.0, 1.0, 0.0));
-
-        let expected = tuple::Vector::new(0.0, 1.0, 0.0);
-        assert_eq!(expected, normal);
-    }
-
-    #[test]
-    fn test_normal_on_a_sphere_on_the_z() {
-        let sphere = sphere::Sphere::default();
-
-        let normal = sphere.normal_at(tuple::Point::new(0.0, 0.0, 1.0));
-
-        let expected = tuple::Vector::new(0.0, 0.0, 1.0);
-        assert_eq!(expected, normal);
-    }
-
-    #[test]
-    fn test_normal_on_a_sphere_at_a_nonaxial_point() {
-        let sphere = sphere::Sphere::default();
-
-        let normal = sphere.normal_at(tuple::Point::new(
-            3.0_f64.sqrt() / 3.0,
-            3.0_f64.sqrt() / 3.0,
-            3.0_f64.sqrt() / 3.0,
-        ));
-
-        let expected = tuple::Vector::new(
-            3.0_f64.sqrt() / 3.0,
-            3.0_f64.sqrt() / 3.0,
-            3.0_f64.sqrt() / 3.0,
-        );
-        assert_eq!(expected, normal);
-    }
-
-    #[test]
-    fn test_normal_is_nornalized_vector() {
-        let sphere = sphere::Sphere::default();
-
-        let normal = sphere.normal_at(tuple::Point::new(
-            3.0_f64.sqrt() / 3.0,
-            3.0_f64.sqrt() / 3.0,
-            3.0_f64.sqrt() / 3.0,
-        ));
-
-        assert_eq!(tuple::normalize(&normal), normal);
-    }
-
-    #[test]
-    fn test_normal_on_a_translated_sphere() {
-        let mut sphere = sphere::Sphere::default();
-        sphere.set_transformation_matrix(matrix::Matrix4::IDENTITY.translation(0.0, 1.0, 0.0));
-
-        let normal = sphere.normal_at(tuple::Point::new(0.0, 1.707107, -0.707107));
-
-        let expected = tuple::Vector::new(0.0, 0.707107, -0.707107);
-        assert_tuple_approx_eq!(expected, normal);
-    }
-
-    #[test]
-    fn test_normal_on_a_transformed_sphere() {
-        let mut sphere = sphere::Sphere::default();
-        sphere.set_transformation_matrix(
-            matrix::Matrix4::IDENTITY
-                .rotation_x(std::f64::consts::PI / 5.0)
-                .scaling(1.0, 0.5, 1.0),
-        );
-
-        let normal = sphere.normal_at(tuple::Point::new(
-            0.0,
-            2.0_f64.sqrt() / 2.0,
-            -2.0_f64.sqrt() / 2.0,
-        ));
-
-        let expected = tuple::Vector::new(0.0, 0.97014, -0.24254);
-        assert_tuple_approx_eq!(expected, normal);
-    }
-
-    #[test]
-    fn test_sphere_has_a_default_material() {
-        let sphere = sphere::Sphere::default();
-        assert_eq!(sphere.material, material::material());
-    }
-
-    #[test]
-    fn test_spheres_material_can_be_set() {
-        let mut sphere = sphere::Sphere::default();
-        let mut material1 = material::material();
-        material1.ambient = 1.0;
-
-        sphere.material = material1;
-
-        assert_eq!(sphere.material.ambient, 1.0);
-    }
-}
