+++ tests/simple_circle_test.rs
@@ -1,8 +1,7 @@
 extern crate ray_tracer;
 use std::io::{Read, Seek};
 
-use ray_tracer::shape::Shape;
-use ray_tracer::{canvas, color, ray, sphere, tuple};
+use ray_tracer::{canvas, color, ray, shape, tuple};
 
 #[test]
 fn test_simple_circle_test() -> Result<(), std::io::Error> {
@@ -13,7 +12,7 @@ fn test_simple_circle_test() -> Result<(), std::io::Error> {
     let canvas_size = 100;
     let pixel_size = wall_size / canvas_size as f64;
     let mut canvas = canvas::canvas(canvas_size, canvas_size);
-    let sphere = sphere::Sphere::default();
+    let sphere = shape::Shape::default_sphere();
 
     for y in 0..canvas.height {
         // Top +half, bottom -half
